requirements.txt
# Core UI Framework
customtkinter==5.2.2

# Image Processing
opencv-python==4.11.0.86
Pillow==11.2.1

# Scientific Computing
numpy==2.0.2

# Face Recognition
insightface==0.7.3
onnx==1.18.0
scikit-learn==1.6.1
scikit-image==0.24.0
protobuf==6.31.1
easydict==1.13
tqdm==4.67.1
PyYAML==6.0.2

# Network scanning and HTTP Requests
python-nmap==0.7.1
requests==2.32.3

# Build Tools
pyinstaller==6.14.1
